Rename graph to visualize and call it from main
