+++ src/main.ts
@@ -2,6 +2,7 @@ import * as fse from 'fs-extra';
 import * as LSIF from 'lsif-protocol';
 import * as yargs from 'yargs';
 import { validate } from './validate';
+import { visualize } from './visualize';
 
 function getInput(format: any, path: string): LSIF.Element[] {
     switch (format) {
@@ -21,6 +22,14 @@ export async function main() {
         })
     }, argv => {
         validate(getInput(argv.inputFormat, argv.file), []);
+    })
+    .command('visualize [file]', '', yargs => {
+        return yargs.positional('file', {
+            describe: 'input file',
+            default: './lsif.json'
+        })
+    }, argv => {
+        visualize(getInput(argv.inputFormat, argv.file), []);
     })
     .option('inputFormat', { default: 'line', choices: ['line', 'json'], description: 'Specify input format' })
     .argv;
